server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.26.0
aiosmtplib==3.0.1
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2
